requirements.txt
langchain
openai
python-dotenv
psycopg2